+++ examples/advanced/src/gen/msw/petController/deletePetHandler.ts
@@ -1,7 +1,9 @@
 import type { DeletePetMutationResponse } from '../../models/ts/petController/DeletePet.ts'
 import { http } from 'msw'
 
-export function deletePetHandler(data?: DeletePetMutationResponse | ((info: Parameters<Parameters<typeof http.delete>[1]>[0]) => Response)) {
+export function deletePetHandler(
+  data?: DeletePetMutationResponse | ((info: Parameters<Parameters<typeof http.delete>[1]>[0]) => Response | Promise<Response>),
+) {
   return http.delete('/pet/:petId\\\\:search', function handler(info) {
     if (typeof data === 'function') return data(info)