+++ examples/advanced/src/gen/msw/petController/uploadFileHandler.ts
@@ -1,7 +1,9 @@
 import type { UploadFileMutationResponse } from '../../models/ts/petController/UploadFile.ts'
 import { http } from 'msw'
 
-export function uploadFileHandler(data?: UploadFileMutationResponse | ((info: Parameters<Parameters<typeof http.post>[1]>[0]) => Response)) {
+export function uploadFileHandler(
+  data?: UploadFileMutationResponse | ((info: Parameters<Parameters<typeof http.post>[1]>[0]) => Response | Promise<Response>),
+) {
   return http.post('/pet/:petId/uploadImage', function handler(info) {
     if (typeof data === 'function') return data(info)