+++ examples/advanced/src/gen/msw/userController/createUsersWithListInputHandler.ts
@@ -2,7 +2,7 @@ import type { CreateUsersWithListInputMutationResponse } from '../../models/ts/u
 import { http } from 'msw'
 
 export function createUsersWithListInputHandler(
-  data?: CreateUsersWithListInputMutationResponse | ((info: Parameters<Parameters<typeof http.post>[1]>[0]) => Response),
+  data?: CreateUsersWithListInputMutationResponse | ((info: Parameters<Parameters<typeof http.post>[1]>[0]) => Response | Promise<Response>),
 ) {
   return http.post('/user/createWithList', function handler(info) {
     if (typeof data === 'function') return data(info)