+++ examples/advanced/src/gen/msw/userController/loginUserHandler.ts
@@ -1,7 +1,7 @@
 import type { LoginUserQueryResponse } from '../../models/ts/userController/LoginUser.ts'
 import { http } from 'msw'
 
-export function loginUserHandler(data?: LoginUserQueryResponse | ((info: Parameters<Parameters<typeof http.get>[1]>[0]) => Response)) {
+export function loginUserHandler(data?: LoginUserQueryResponse | ((info: Parameters<Parameters<typeof http.get>[1]>[0]) => Response | Promise<Response>)) {
   return http.get('/user/login', function handler(info) {
     if (typeof data === 'function') return data(info)