+++ examples/msw/src/gen/msw/pet/Handlers/deletePetHandler.ts
@@ -6,7 +6,9 @@
 import type { DeletePetMutationResponse } from '../../../models/DeletePet.ts'
 import { http } from 'msw'
 
-export function deletePetHandler(data?: DeletePetMutationResponse | ((info: Parameters<Parameters<typeof http.delete>[1]>[0]) => Response)) {
+export function deletePetHandler(
+  data?: DeletePetMutationResponse | ((info: Parameters<Parameters<typeof http.delete>[1]>[0]) => Response | Promise<Response>),
+) {
   return http.delete('http://localhost:3000/pet/:petId\\\\:search', function handler(info) {
     if (typeof data === 'function') return data(info)