+++ examples/msw/src/gen/msw/pet/Handlers/getPetByIdHandler.ts
@@ -6,7 +6,7 @@
 import type { GetPetByIdQueryResponse } from '../../../models/GetPetById.ts'
 import { http } from 'msw'
 
-export function getPetByIdHandler(data?: GetPetByIdQueryResponse | ((info: Parameters<Parameters<typeof http.get>[1]>[0]) => Response)) {
+export function getPetByIdHandler(data?: GetPetByIdQueryResponse | ((info: Parameters<Parameters<typeof http.get>[1]>[0]) => Response | Promise<Response>)) {
   return http.get('http://localhost:3000/pet/:petId\\\\:search', function handler(info) {
     if (typeof data === 'function') return data(info)