+++ examples/msw/src/gen/msw/store/Handlers/deleteOrderHandler.ts
@@ -6,7 +6,9 @@
 import type { DeleteOrderMutationResponse } from '../../../models/DeleteOrder.ts'
 import { http } from 'msw'
 
-export function deleteOrderHandler(data?: DeleteOrderMutationResponse | ((info: Parameters<Parameters<typeof http.delete>[1]>[0]) => Response)) {
+export function deleteOrderHandler(
+  data?: DeleteOrderMutationResponse | ((info: Parameters<Parameters<typeof http.delete>[1]>[0]) => Response | Promise<Response>),
+) {
   return http.delete('http://localhost:3000/store/order/:orderId', function handler(info) {
     if (typeof data === 'function') return data(info)