+++ examples/msw/src/gen/msw/user/Handlers/createUserHandler.ts
@@ -6,7 +6,9 @@
 import type { CreateUserMutationResponse } from '../../../models/CreateUser.ts'
 import { http } from 'msw'
 
-export function createUserHandler(data?: CreateUserMutationResponse | ((info: Parameters<Parameters<typeof http.post>[1]>[0]) => Response)) {
+export function createUserHandler(
+  data?: CreateUserMutationResponse | ((info: Parameters<Parameters<typeof http.post>[1]>[0]) => Response | Promise<Response>),
+) {
   return http.post('http://localhost:3000/user', function handler(info) {
     if (typeof data === 'function') return data(info)