+++ examples/msw/src/gen/msw/user/Handlers/createUsersWithListInputHandler.ts
@@ -7,7 +7,7 @@ import type { CreateUsersWithListInputMutationResponse } from '../../../models/C
 import { http } from 'msw'
 
 export function createUsersWithListInputHandler(
-  data?: CreateUsersWithListInputMutationResponse | ((info: Parameters<Parameters<typeof http.post>[1]>[0]) => Response),
+  data?: CreateUsersWithListInputMutationResponse | ((info: Parameters<Parameters<typeof http.post>[1]>[0]) => Response | Promise<Response>),
 ) {
   return http.post('http://localhost:3000/user/createWithList', function handler(info) {
     if (typeof data === 'function') return data(info)