+++ examples/msw/src/gen/msw/user/Handlers/deleteUserHandler.ts
@@ -6,7 +6,9 @@
 import type { DeleteUserMutationResponse } from '../../../models/DeleteUser.ts'
 import { http } from 'msw'
 
-export function deleteUserHandler(data?: DeleteUserMutationResponse | ((info: Parameters<Parameters<typeof http.delete>[1]>[0]) => Response)) {
+export function deleteUserHandler(
+  data?: DeleteUserMutationResponse | ((info: Parameters<Parameters<typeof http.delete>[1]>[0]) => Response | Promise<Response>),
+) {
   return http.delete('http://localhost:3000/user/:username', function handler(info) {
     if (typeof data === 'function') return data(info)