+++ packages/plugin-msw/src/generators/__snapshots__/createPetFaker.ts
@@ -4,7 +4,7 @@
  */
 import { http } from 'msw'
 
-export function createPets(data?: CreatePetsMutationResponse | ((info: Parameters<Parameters<typeof http.post>[1]>[0]) => Response)) {
+export function createPets(data?: CreatePetsMutationResponse | ((info: Parameters<Parameters<typeof http.post>[1]>[0]) => Response | Promise<Response>)) {
   return http.post('/pets', function handler(info) {
     if (typeof data === 'function') return data(info)