+++ packages/plugin-msw/src/generators/__snapshots__/deletePet.ts
@@ -4,7 +4,9 @@
  */
 import { http } from 'msw'
 
-export function deletePetsPetid(data?: DeletePetsPetidMutationResponse | ((info: Parameters<Parameters<typeof http.delete>[1]>[0]) => Response)) {
+export function deletePetsPetid(
+  data?: DeletePetsPetidMutationResponse | ((info: Parameters<Parameters<typeof http.delete>[1]>[0]) => Response | Promise<Response>),
+) {
   return http.delete('/pets/:petId', function handler(info) {
     if (typeof data === 'function') return data(info)