+++ packages/plugin-client/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @kubb/plugin-client
 
+## 4.1.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @kubb/core@4.1.3
+  - @kubb/oas@4.1.3
+  - @kubb/plugin-oas@4.1.3
+  - @kubb/plugin-ts@4.1.3
+  - @kubb/plugin-zod@4.1.3
+  - @kubb/react@4.1.3
+
 ## 4.1.2
 
 ### Patch Changes